test(filters): move generic filter tests out of test_automations
Relocate the generic MongoDB filter tests to a new tests/unit_tests/test_filters/
package that mirrors the relocated wandb/_filters/ code. Split the generic filter
hypothesis strategies into a co-located _strategies.py and have the automations
tests import the shared generic helpers from there.
