+++ tests/unit_tests/test_automations/test_run_events.py
@@ -20,6 +20,7 @@ from wandb.automations._run_metric_filters import Agg, ChangeDir, MetricAgg, Met
 from wandb.automations._run_state_filters import ReportedRunState
 from wandb.automations.events import StateFilter
 
+from ..test_filters._strategies import ints_or_floats
 from ._automation_strategies import (
     aggs,
     cmp_keys,
@@ -32,7 +33,6 @@ from ._automation_strategies import (
     run_states,
     window_sizes,
 )
-from ._strategies import ints_or_floats
 
 
 @composite
